requirements.txt
aiokafka==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.1.3
executing==2.0.1
greenlet==3.0.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kafka-python==2.0.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
platformdirs==4.2.0
playwright==1.39.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
py4j==0.10.9.7
pyee==11.0.1
Pygments==2.17.2
pymongo==4.6.2
pyspark==3.5.1
psycopg2==2.9.9
pytest==7.4.3
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2024.1
pywin32==306
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
text-unidecode==1.3
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.13
